+++ pkg/acceptance/bettertestspoc/config/datasourcemodel/streamlits_model_ext.go
@@ -6,10 +6,26 @@ import (
 	tfconfig "github.com/hashicorp/terraform-plugin-testing/config"
 )
 
+func (s *StreamlitsModel) WithInAccount() *StreamlitsModel {
+	return s.WithInValue(
+		tfconfig.ObjectVariable(map[string]tfconfig.Variable{
+			"account": tfconfig.BoolVariable(true),
+		}),
+	)
+}
+
 func (s *StreamlitsModel) WithInDatabase(databaseId sdk.AccountObjectIdentifier) *StreamlitsModel {
 	return s.WithInValue(
 		tfconfig.ObjectVariable(map[string]tfconfig.Variable{
 			"database": tfconfig.StringVariable(databaseId.Name()),
+		}),
+	)
+}
+
+func (s *StreamlitsModel) WithInSchema(schemaId sdk.DatabaseObjectIdentifier) *StreamlitsModel {
+	return s.WithInValue(
+		tfconfig.ObjectVariable(map[string]tfconfig.Variable{
+			"schema": tfconfig.StringVariable(schemaId.FullyQualifiedName()),
 		}),
 	)
 }