+++ pkg/sdk/generator/example/defs/0_init.go
@@ -1,0 +1,21 @@
+//go:build sdk_generation_examples
+
+package defs
+
+import (
+	"fmt"
+
+	"github.com/Snowflake-Labs/terraform-provider-snowflake/pkg/sdk/generator/gen"
+)
+
+func init() {
+	gen.AllSdkObjectDefinitions = append(gen.AllSdkObjectDefinitions,
+		DatabaseRoles,
+		ToOptsOptionalExample,
+		SequencesDef,
+	)
+	fmt.Println("SDK object definitions:")
+	for _, def := range gen.AllSdkObjectDefinitions {
+		fmt.Printf(" - %s\n", def.Name)
+	}
+}