+++ pkg/sdk/grants_validations.go
@@ -54,6 +54,55 @@ var validGrantOwnershipObjectTypes = []ObjectType{
 	ObjectTypeSchema,
 	ObjectTypeSessionPolicy,
 	ObjectTypeSecret,
+	ObjectTypeSemanticView,
+	ObjectTypeSequence,
+	ObjectTypeStage,
+	ObjectTypeStream,
+	ObjectTypeTable,
+	ObjectTypeTag,
+	ObjectTypeTask,
+	ObjectTypeUser,
+	ObjectTypeView,
+	ObjectTypeWarehouse,
+}
+
+// Database roles are excluded
+var validGrantOwnershipBulkObjectTypes = []ObjectType{
+	ObjectTypeAggregationPolicy,
+	ObjectTypeAlert,
+	ObjectTypeAuthenticationPolicy,
+	ObjectTypeComputePool,
+	ObjectTypeDataMetricFunction,
+	ObjectTypeDatabase,
+	ObjectTypeDynamicTable,
+	ObjectTypeEventTable,
+	ObjectTypeExternalTable,
+	ObjectTypeExternalVolume,
+	ObjectTypeFailoverGroup,
+	ObjectTypeFileFormat,
+	ObjectTypeFunction,
+	ObjectTypeGitRepository,
+	ObjectTypeHybridTable,
+	ObjectTypeIcebergTable,
+	ObjectTypeImageRepository,
+	ObjectTypeIntegration,
+	ObjectTypeMaterializedView,
+	ObjectTypeNetworkPolicy,
+	ObjectTypeNetworkRule,
+	ObjectTypePackagesPolicy,
+	ObjectTypePipe,
+	ObjectTypeProcedure,
+	ObjectTypeMaskingPolicy,
+	ObjectTypePasswordPolicy,
+	ObjectTypeProjectionPolicy,
+	ObjectTypeReplicationGroup,
+	ObjectTypeResourceMonitor,
+	ObjectTypeRole,
+	ObjectTypeRowAccessPolicy,
+	ObjectTypeSchema,
+	ObjectTypeSessionPolicy,
+	ObjectTypeSecret,
+	ObjectTypeSemanticView,
 	ObjectTypeSequence,
 	ObjectTypeStage,
 	ObjectTypeStream,
@@ -72,6 +121,7 @@ var validGrantToAccountObjectTypes = []ObjectType{
 	ObjectTypeComputePool,
 	ObjectTypeDatabase,
 	ObjectTypeIntegration,
+	ObjectTypeConnection,
 	ObjectTypeFailoverGroup,
 	ObjectTypeReplicationGroup,
 	ObjectTypeExternalVolume,
@@ -114,22 +164,45 @@ var validGrantToSchemaObjectTypes = []ObjectType{
 	ObjectTypeService,
 	ObjectTypeSessionPolicy,
 	ObjectTypeSequence,
+	ObjectTypeStorageLifecyclePolicy,
 	ObjectTypeSnapshot,
 	ObjectTypeSnapshotPolicy,
 	ObjectTypeSnapshotSet,
 	ObjectTypeStage,
 	ObjectTypeStream,
 	ObjectTypeStreamlit,
+	ObjectTypeOnlineFeatureTable,
 	ObjectTypeTable,
 	ObjectTypeTag,
 	ObjectTypeTask,
 	ObjectTypeView,
+	ObjectTypeWorkspace,
+}
+
+// TODO(SNOW-2370066): Adjust after adding tests
+// based on https://docs.snowflake.com/en/sql-reference/sql/grant-privilege#restrictions-and-limitations
+var invalidGrantToAllObjectTypes = []ObjectType{
+	ObjectTypeComputePool,
+	ObjectTypeExternalFunction,
+	ObjectTypeExperiment,
+	// ObjectTypeImageRepository,
+	// ObjectTypeAggregationPolicy,
+	// ObjectTypeMaskingPolicy,
+	// ObjectTypePackagesPolicy,
+	// ObjectTypeProjectionPolicy,
+	// ObjectTypeRowAccessPolicy,
+	// ObjectTypeSessionPolicy,
+	ObjectTypeStorageLifecyclePolicy,
+	// ObjectTypeTag,
+	ObjectTypeWarehouse,
+	ObjectTypeWorkspace,
 }
 
 // based on https://docs.snowflake.com/en/sql-reference/sql/grant-privilege#restrictions-and-limitations
 var invalidGrantToFutureObjectTypes = []ObjectType{
 	ObjectTypeComputePool,
 	ObjectTypeExternalFunction,
+	ObjectTypeExperiment,
 	ObjectTypeImageRepository,
 	ObjectTypeAggregationPolicy,
 	ObjectTypeMaskingPolicy,
@@ -137,21 +210,26 @@ var invalidGrantToFutureObjectTypes = []ObjectType{
 	ObjectTypeProjectionPolicy,
 	ObjectTypeRowAccessPolicy,
 	ObjectTypeSessionPolicy,
+	ObjectTypeStorageLifecyclePolicy,
 	ObjectTypeTag,
+	ObjectTypeWarehouse,
+	ObjectTypeWorkspace,
 }
 
 var (
 	ValidGrantOwnershipObjectTypesString       = make([]string, len(validGrantOwnershipObjectTypes))
-	ValidGrantOwnershipPluralObjectTypesString = make([]string, len(validGrantOwnershipObjectTypes))
+	ValidGrantOwnershipPluralObjectTypesString = make([]string, len(validGrantOwnershipBulkObjectTypes))
 	ValidGrantToAccountObjectTypesString       = make([]string, len(validGrantToAccountObjectTypes))
 	ValidGrantToSchemaObjectTypesString        = make([]string, len(validGrantToSchemaObjectTypes))
-	ValidGrantToPluralObjectTypesString        = make([]string, len(validGrantToSchemaObjectTypes))
+	ValidGrantToAllPluralObjectTypesString     = make([]string, 0)
 	ValidGrantToFuturePluralObjectTypesString  = make([]string, 0)
 )
 
 func init() {
 	for i, objectType := range validGrantOwnershipObjectTypes {
 		ValidGrantOwnershipObjectTypesString[i] = objectType.String()
+	}
+	for i, objectType := range validGrantOwnershipBulkObjectTypes {
 		ValidGrantOwnershipPluralObjectTypesString[i] = objectType.Plural().String()
 	}
 	for i, objectType := range validGrantToAccountObjectTypes {
@@ -159,7 +237,9 @@ func init() {
 	}
 	for i, objectType := range validGrantToSchemaObjectTypes {
 		ValidGrantToSchemaObjectTypesString[i] = objectType.String()
-		ValidGrantToPluralObjectTypesString[i] = objectType.Plural().String()
+		if !slices.Contains(invalidGrantToAllObjectTypes, objectType) {
+			ValidGrantToAllPluralObjectTypesString = append(ValidGrantToAllPluralObjectTypesString, objectType.Plural().String())
+		}
 		if !slices.Contains(invalidGrantToFutureObjectTypes, objectType) {
 			ValidGrantToFuturePluralObjectTypesString = append(ValidGrantToFuturePluralObjectTypesString, objectType.Plural().String())
 		}